consolidate react context providers into a single component called AppProvider
