+++ client/app/App.tsx
@@ -1,37 +1,15 @@
-import { ReactElement, Suspense } from 'react';
-import { Container } from 'react-bootstrap';
+import { ReactElement } from 'react';
 import { RouterProvider } from 'react-router-dom';
-import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
-import { ErrorBoundary } from '~/components/Error';
-import { Notifications } from '~/components/Notifications/Notifications';
+import { AppProvider } from '~/providers';
 import { router } from '~/routes';
-import { Spinner } from './components/ui';
 import './globals.css';
-
-const GlobalSpinner = () => (
-  <Container fluid className="d-flex justify-content-center align-items-center vh-100">
-    <Spinner size="sm" className="my-auto" />
-  </Container>
-);
 
 function App(): ReactElement {
   return (
-    <ErrorBoundary>
-      <ToastContainer
-        position="bottom-right"
-        autoClose={5000}
-        hideProgressBar={false}
-        newestOnTop
-        closeOnClick
-        pauseOnHover
-        limit={3}
-      />
-      <Notifications />
-      <Suspense fallback={<GlobalSpinner />}>
-        <RouterProvider router={router} />
-      </Suspense>
-    </ErrorBoundary>
+    <AppProvider>
+      <RouterProvider router={router} />
+    </AppProvider>
   );
 }
