+++ client/app/index.tsx
@@ -1,8 +1,6 @@
-import { createRoot } from 'react-dom/client';
-import { Provider } from 'react-redux';
-import { rootStore } from '~/store';
-import App from './App';
 import React from 'react';
+import { createRoot } from 'react-dom/client';
+import App from './App';
 
 // Start mock service worker in development mode
 async function enableMocking() {
@@ -19,9 +17,7 @@ const root = createRoot(container);
 enableMocking().then(() => {
   root.render(
     <React.StrictMode>
-      <Provider store={rootStore}>
-        <App />
-      </Provider>
+      <App />
     </React.StrictMode>
   );
 });
